Add --server to the run script to install environment and run server, with no test

+++ tests/run_java_tests.sh
@@ -5,19 +5,36 @@ type nc >/dev/null 2>&1 || { echo >&2 "Netcat required but it's not installed.  
 
 mkdir -p ./env_java_tests/lib/ 2> /dev/null
 
-if [ ! -d ./env_java_tests/lib/spark-example/ ]; then
-  rm  -rf ./env_java_tests/lib/spark-app/
-  cp -rf ./env_java_tests/spark-app ./env_java_tests/lib/spark-app/
+GRADLEPID=0
+
+# Run  webserver
+function run_webserver()
+{
+
+  if [ ! -d ./env_java_tests/lib/spark-example/ ]; then
+    rm  -rf ./env_java_tests/lib/spark-app/
+    cp -rf ./env_java_tests/spark-app ./env_java_tests/lib/spark-app/
+  fi
+
+  cd ./env_java_tests/lib/spark-app/
+  gradle run
+  GRADLEPID=$!
+  cd ../../../
+}
+
+
+if [[ "$1" == "--server" ]]; then
+  echo 'Raise web server only'
+  run_webserver
+else
+  echo 'Run web server and launch tests'
+  run_webserver &
+
+  while ! echo exit | nc localhost 15003; do sleep 3; done
+
+
+  python -m unittest discover . 'test_java_*.py'
+
+  # Shutdown Java webserver
+  kill ${GRADLEPID}
 fi
-
-cd ./env_java_tests/lib/spark-app/
-gradle run &
-GRADLEPID=$!
-
-while ! echo exit | nc localhost 15003; do sleep 3; done
-
-cd ../../../
-python -m unittest discover . 'test_java_*.py'
-sleep 1
-# Shutdown Java webserver
-kill ${GRADLEPID}
